switch all docs to roxygen
This is a first stab.  I am using templates for quite a few things but
probably not all.  I generally favour using templates, because then if I
discover that the wording can be improved, I only need to update one
file not a dozen files.

With these (and other recent) changes, it took me a little while to get
the whole thing to build and check, but it does now, and that's why I am
commiting the code now.